Title: 선두를 지켜라 3, Time: 17ms, Memory: 0MB, Status: Runtime Error - Codetree

+++ 250715/선두를 지켜라 3/keep-the-lead-3.cpp
@@ -7,7 +7,7 @@ int main()
     int N, M;
     cin >> N >> M;
 
-    int A[1001];
+    unsigned long long A[1001];
     A[0] = 0;
     int AIndex = 1;
     int totalTime = 0;
@@ -26,7 +26,7 @@ int main()
         totalTime += t;
     }
 
-    int B[1001];
+    unsigned long long B[1001];
     B[0] = 0;
     int BIndex = 1;
